LayoutTests/fast/events/event-constructors.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description("This tests the constructors for all the event DOM classes that have them.");

shouldThrow("new Event()");
shouldBe("new Event(null).type", "'null'");
shouldBe("new Event(undefined).type", "'undefined'");
shouldBe("new Event('foo').type", "'foo'");
shouldBe("new Event('foo', {}).type", "'foo'");
shouldBe("new Event('foo', {'bubbles': false}).bubbles", "false");
shouldBe("new Event('foo', {'bubbles': true}).bubbles", "true");
shouldBe("new Event('foo', {'cancelable': false}).cancelable", "false");
shouldBe("new Event('foo', {'cancelable': true}).cancelable", "true");
shouldBe("new Event('foo', {'defaultPrevented': false}).defaultPrevented", "false");
shouldBe("new Event('foo', {'defaultPrevented': true}).defaultPrevented", "true");
shouldBe("new Event('foo', {'timeStamp': 1234}).timeStamp", "1234");

var successfullyParsed = true;
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
