LayoutTests/fast/events/event-constructors-expected.txt

This tests the constructors for all the event DOM classes that have them.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new Event() threw exception TypeError: Not enough arguments.
PASS new Event(null).type is 'null'
PASS new Event(undefined).type is 'undefined'
PASS new Event('foo').type is 'foo'
PASS new Event('foo', {}).type is 'foo'
PASS new Event('foo', {'bubbles': false}).bubbles is false
PASS new Event('foo', {'bubbles': true}).bubbles is true
PASS new Event('foo', {'cancelable': false}).cancelable is false
PASS new Event('foo', {'cancelable': true}).cancelable is true
PASS new Event('foo', {'defaultPrevented': false}).defaultPrevented is false
PASS new Event('foo', {'defaultPrevented': true}).defaultPrevented is true
PASS new Event('foo', {'timeStamp': 1234}).timeStamp is 1234
PASS successfullyParsed is true

TEST COMPLETE
